Fix read state observation crash
Using KVO with Core Data seems to be a minefield. Removing observers in
`willTurnIntoFault()` was causing a strange crash, and I'm not
interested in debugging managed object lifetime issues. Switching to
`NotificationCenter` gives two advantages: we get automatic observer
removal, and observation of changed objects at the granularity that's
most useful for aggregating the read state of plan chapters.